Linkedin issue
I've addressed an issue where the LinkedIn information was still appearing on the resume templates even if the user did not input their LinkedIn URL. The solution involved modifying the JavaScript code responsible for updating the CV preview. Specifically, I added a condition to check whether the LinkedIn input field is empty and, based on that, decide whether to display the LinkedIn information in the templates. This was done by adjusting the updateCVPreview(templateName) function. Now, the templates will only show the LinkedIn section if the user has provided their LinkedIn URL, making sure that no placeholder or irrelevant information is displayed on the final resume.